% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{cr_isbn_lookup}
\alias{cr_isbn_lookup}
\title{Crossref Works given an ISBN}
\usage{
cr_isbn_lookup(isbn)
}
\arguments{
\item{isbn}{character the ISBN to look up information for}
}
\value{
data frame with information
}
\description{
Crossref Works API provides information given an ISBN
and here we return the title, DOIs and related ISBNs.
}
\details{
For more information, see \url{# https://api.crossref.org/swagger-ui/index.html#/Works/get_works}
}
\examples{
\dontrun{
if(interactive()){
 cr_isbn_lookup("9783030755379")
 }
}
}
