% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{cr_name_parts}
\alias{cr_name_parts}
\title{Crossref information about name parts, given DOI(s)}
\usage{
cr_name_parts(doi)
}
\arguments{
\item{doi}{character vector w DOI(s) to look up information for}
}
\value{
data frame with name parts info
}
\description{
Crossref Works API provides information about name parts
for some DOIs.
}
\details{
For more information, see \url{#https://api.crossref.org/swagger-ui/index.html#/Works/get_works__doi_}
}
\examples{
\dontrun{
if(interactive()){
 cr_name_parts("10.1021/cr900356p")
 }
}
}
