% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr24.R
\name{fetch_hr24}
\alias{fetch_hr24}
\title{Fetch latest HR file}
\usage{
fetch_hr24(bucket = "hr24", filedate = Sys.Date())
}
\arguments{
\item{bucket}{The bucket to look in for the file, default hr24}

\item{filedate}{an optional date to look for, default Sys.Date}
}
\description{
Fetch latest HR file
}
