% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthcorpus-package.R
\docType{data}
\name{hr24_employment_title}
\alias{hr24_employment_title}
\title{HR24 (Visma) lookup table}
\format{
A data frame with 65 rows and 4 variables:
\describe{
\item{\code{emp_code}}{character the 5-digit employee code}
\item{\code{emp_desc}}{character role description in Swedish}
\item{\code{title_sv}}{character title in Swedish}
\item{\code{title_en}}{character title in English}
\item{\code{cat_desc}}{character category description in Swedish}
\item{\code{is_uf_ta}}{character notes if role is UF or TA}
}
}
\usage{
hr24_employment_title
}
\description{
A lookup table for employee codes and corresponding employee descriptions
and titles
}
\details{
DETAILS
}
\keyword{datasets}
