% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthcorpus-package.R
\docType{data}
\name{scb_topics}
\alias{scb_topics}
\title{SCB/UKÄ research topics}
\format{
A data frame with 345 rows and 6 variables:
\describe{
\item{\code{id}}{integer An id with 1, 3 or 5 digits}
\item{\code{Swedish}}{character Swedish name of topic}
\item{\code{English}}{character English name of topic}
\item{\code{level}}{integer Level of topic 1-3}
\item{\code{Swe_comment}}{character Extra information in Swedish if available}
\item{\code{Eng_comment}}{character Extra information in English if available}
}
}
\usage{
scb_topics
}
\description{
Research topics from UKÄ/SCB, 2025 version
}
\keyword{datasets}
